refactor: split recorder component
